requirements.txt
Django==4.2
pillow==10.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-allauth==0.57.0